Fix support for TFTrue
The rgl configs were changing the default log file location. TFTrue is
not smart enough to look for logs in this new location so every single
log was failing to upload. Commenting out the relevant line in the rgl
configs was enough to get TFTrue working again and I can now upload logs
like normal.